extract CorsBundle
The code to set up the CrossOriginFilter is pretty generic, in that it
could easily be applied to other projects without modification.  We
can extract this into a Bundle and try to keep the RegisterApplication
smaller.

I hardcoded the specific CORS configuration values for the moment.  If
we ever do extract CorsBundle into a library, we can turn it into a
ConfiguredBundle at that point.