Don't baseline flyway any more
When we first rolled out flyway, we needed to use the baseline
functionality to adapt flyway to the existing data schemas we had.

However, now that everything uses flyway for schema management, we
don't need to enable this.

Furthermore, all future registers we create will start from an empty
schema, so won't need the baselining functionality.

This is possible because our V1 script is idempotent - it uses CREATE
TABLE IF NOT EXISTS commands to ensure it worked on our existing
estate but can also be used on new databases.