Change "custodian" to "registry"
As the register history pages show, the "custodian" of a register is a
named individual; while the organisation the custodian belongs to is
known as a "registry".  Our code should use this same language.

I've tested this locally to make sure the templates continue to get
the custodian/registry branding stuff correctly.