Use RegisterReadOnly interface, not RegisterData, in views
This migrates the views to use the RegisterReadOnly interface instead
of RegisterData.  It also adds getRegisterName() and
getRegisterMetadata() to that interface.

My idea here is that we can start to migrate other things to the
RegisterReadOnly interface - like getting the custodian and the
branding.

This also will help with the work I'm doing towards multiple registers
in one app server.