Change the way we manage Record "list" views
Same deal, but for "record lists" -- actually, they're objects when
viewed as JSON, which is why I renamed "RecordListView" to
"RecordsView".

When I made this commit I discovered a slight wrinkle with splitting
up the resource methods between different media types like this:
whereas previously, Jersey would pick the first mediatype in the
@Produces annotation as the default, now it will pick the first
mediatype in a randomly-selected method. This means that if you
request `/item/sha-256:fa1afe1...abcde` with the header `Accept: */*`,
you might end up with JSON instead of HTML.

This affects cURL, and some of our tests, but doesn't affect browsers
because they specify `Accept: text/html` (well, something a bit more
complicated than this but near enough the same). I suspect this is
good enough?