commit TransactionalMemoizationStore changes back after txn
Once the transaction has committed, it is safe to commit the hashes
back to the store.

It's possible that something will fail between the DB transaction
commit and the TransactionalMemoizationStore commit.  That's okay,
because all it means is that we'll have to recompute some hashes, not
that any incorrect behaviour will occur.