Move registerName into ItemValidator
There was a weird pattern here where ItemValidator needed to be told
the registerName on every call to validateItem(), and EntryLog only
had a registerName field in order to pass it on to the ItemValidator.

I moved the field into the ItemValidator and cleaned up.

(Incidentally, I think this is a case of reverse data envy:
http://wiki.c2.com/?DataEnvy -- reverse because EntryLog wants to give
the data to ItemValidator rather than wants to get some bit of data
out of ItemValidator)