consistent naming - totalEntries1/2 not start/end
This iterator method returns an iterator of entries that occured
between the register at two different sizes.  So
entryQueryDAO.getIterator(5,10) should return entries 6,7,8,9 and 10.

The "start"/"end" naming is confusing because it's not clear what the
boundary conditions should be.  By naming in terms of total entries,
it becomes clearer again.