create RegisterName class
This creates a RegisterName class, which is mostly a thin wrapper
around String for those Strings which specifically name registers.
This achieves a couple of things:

 - a RegisterName can more easily be injected on its own
 - a String can't be passed where a RegisterName is expected, and vice
   versa

This also reduces some duplications: for example, between
AssetsBundleCustomErrorHandler and ThymeleafView (which were both
constructing friendly names) and between
AssetsBundleCustomErrorHandler and RequestContext (which were both
extracting register names from the Host/X-Forwarded-Host header).