add more tests for TransactionalRecordIndex
In the big refactoring, we lost some tests that verified we were
flushing results out to the database before making queries.  This
restores them in a hacky way -- the tests *only* check that we wrote out
to disk.  Even with this hacky level of testing, they caught a bug -
which is fixed in this commit too :)

It's hard to test TransactionalRecordIndex in the same way that we test
TransactionalItemStore and TransactionalEntryLog.  The EntryLog tests
are easiest because an EntryLog stands alone and has no other
dependencies.  The ItemStore tests are a little harder because an
ItemStore needs to access Entry data as well as Item data, so the
InMemoryItemDAO is more complex.

For RecordIndex, this is even more complex, because the RecordQueryDAO
needs to access data about Records, Items and Entries, so you end up
needing to fake out basically the whole production database schema.

I think the right way to fix this will be to make RecordIndex go away by
turning it into an index on the Entry table :)