Fail to start if we're missing required RegisterData
Currently, on startup we create all the RegisterContext instances we
need, but we don't check that RegistersConfiguration has the required
RegisterData until the first request is routed to that register.

This changes RegisterContext to pull out the needed RegisterData in
its constructor, which will force the app to fail startup if the
RegisterData can't be found.