rename `otherRegisters` to `registers`
Longer-term, I think we're likely to get rid of the default register
configuration items (the top-level things like `database`, `register`,
`enableRegisterDataDelete` etc).  At that point, the name
`otherRegisters` won't make sense any more.

This renames it to plain `registers`.

Nothing is yet consuming this configuration item except the config
files in this repo, so it's a safe rename to do.