Make similar registers a List<>, not an Optional<List<>>
The `Optional<List<String>>` was a bit of a code smell, because it
allows separate values for `Optional.empty()` and
`Optional.of(emptyList())`.

If you had a config file with `similarRegisters: []`, you would render
a page which had a "Similar Registers" heading but with no registers
in the list.

This transforms the `Optional<List<String>>` to a plain
`List<String>`.  If the config item is missing, it defaults to empty
list.  We only render the "Similar Registers" heading if the list is
nonempty.