Add new container to run fluentd, fluentd-sumologic plugin and update openregister to log to fluentd.
Set up a new sidecar container to run fluentd and the fluentd-sumologic plugin (https://github.com/SumoLogic/fluentd-output-sumologic), then update the docker container running openregister to use --log-driver=fluentd.

The docker image being used to run fluentd and the fluentd-sumologic plugin is currently `samcrang/fluentd-sumologic`, as we were having issues with permissions when connecting dockerhub to the github openregisters repo. This is something we are looking to fix.