Refresh register metadata after deleting data
This fixes the update of the fields. If a user modifies register or field
data and then resets a register (eg by calling /delete-register-data)
the register metadata is not being refreshed making home page look the same etc.

The bug was in RegisterContext class which was storing RegisterMetadata
as field passed in the constructor hence after a reload even though
ConfigManager  had never data Context was still using it's field.