add total-entries to register proof
This adds the `total-entries` to the register proof, in line with the
specification after openregister/specification#66.

As a result, there are some places which were using a RegisterProof
but only really needed the root HashValue from inside it -- in
particular, code dealing with RSF.  I've refactored those bits of code
to use HashValue directly instead of RegisterProof.