Time all Dropwizard Resource methods
This means we output some metrics for each resource, for example for
the `/robots.txt` path we'll start generating metrics that look like the
following:

```
uk.gov.register.resources.HomePageResource.robots
             count = 15
         mean rate = 1.21 calls/second
     1-minute rate = 0.05 calls/second
     5-minute rate = 0.01 calls/second
    15-minute rate = 0.00 calls/second
               min = 0.10 milliseconds
               max = 7.64 milliseconds
              mean = 0.61 milliseconds
            stddev = 1.85 milliseconds
            median = 0.12 milliseconds
              75% <= 0.14 milliseconds
              95% <= 7.64 milliseconds
              98% <= 7.64 milliseconds
              99% <= 7.64 milliseconds
            99.9% <= 7.64 milliseconds
```

At the moment this won't provide a per-register time for these
resources. If we want to do that we'll probably need to change the way
Dropwizard behaves when reflecting over methods with the `Timed`
annotation. Probably will have to look at:

```
com.codahale.metrics.jersey2.InstrumentedResourceMethodApplicationListener
```