Use Travis's Postgres addon
Previously, Travis did not provide Postgres 9.5 as an addon. In order to
use this we had to install a version manually (see commit
`e05a0323ec67b65eb2551611c7562afb6d2cd519`).

[Travis now provide Postgres 9.5 as an addon](https://github.com/travis-ci/travis-ci/issues/4264#issuecomment-263550556).

I have a feeling that we should be able to use Travis containers
(setting `sudo: false`) but it appeared to fail silently when I tried
it:

```
$ ./gradlew check
:dependencyCheck
Verifying dependencies for project openregister-java
Checking for updates and analyzing vulnerabilities for dependencies
/home/travis/.travis/job_stages: line 53:  3800 Killed                ./gradlew check
```

The builds also appeared to be slower.