Fix ZDD deploys in discovery environment
There is a max length for the names of ELBs; if we have an ELB with the
name `$ENV--$REGISTER-NAME` then the max length is exceeded in the
discovery environment. We specifically deal with this case by trimming
`discovery` to `disco` in our Terraform configuration. We need to mirror
that trimming here too.