Add script to spin up a register environment locally
This will spin up a new register environment (an individual register as
well as the "basic" registers). The register points to the basic
registers over HTTP. The data loaded is cloned from an existing
environment (by default "alpha") but can be set to something else using
the `ENVIRONMENT` environment variable.