Ensure database has all migrations applied before tests run
This introduces a rule that runs the flyway migrations before
tests are run, which creates all the required database tables.
Some functional tests use the existing RegisterRule, which runs
the migrations as part of starting up the application. For the
EntryMapperTest class however, it does not use RegisterRule and
the migration is not run. If the test database has been newly
created, you would sometimes see problems where address.entry
table did not exist. On occasions where tests using the
RegisterRule run before EntryMapperTest you would not see this
problem as the migration would be performed before EntryMapperTest
is run. I chose to implement a new MigrateDatabaseRule rather than
use the RegisterRule because RegisterRule would have caused
the application to unnecessarily start up and slow down these tests.