Introduce concept of non-existant item primary key
At the moment all register have primary key fields in the item.
This field is always the field with the same name as the register.
It is now possible that indexes will be loaded into
openregister-java and there may not be a field that exists with
the same name as the register. In that case, we want to know that
there does not exist an item primary key.