Redirect HTTP to HTTPS if application is behind a load balancer
If the application is running behind a load balancer and forwards the
original protocol the user used in the `X-Forwarded-Proto` header and
the original protocol is HTTP then upgrade to HTTPS.

This should only exist for as long as we don't have something else
fronting the application that can deal with protocol upgrades. We're
doing this here because it avoids having to create a reverse-proxy (or
similar) which we don't have. It absolutely is not the responsible of
openregister-java to know or care about where it is running.