Make rendering of records and record more consistent for indexes and registers
This change extracts some of html code into a fragment for an
individual item row. If it is a record page it always resolves
links to registers. If it is a records page it does not render any
links to other registers. However, if there is no key column displayed
then the first item column is used to render a link to the "drilled-down"
view. i.e. /record/bla or /index/bla/record/blabla. The code has also
been changed to assume that there may not be an item primary key.