Refactor RecordView and RecordsView
There was a lot of logic duplicated in the RecordView and
RecordsView. This simplifies RecordView to use much of the
logic that exists in RecordsView because it is the same except
that the list of records only contains one thing.
I also changed Record to use a list for items instead of a
Hashmap. This ensures that item ordering in records is now stable.
The items property of the Record object has also been removed from
the equals method because two records should be equal if their
entries are equal. The items property is populated for convenience
when generating views and should not need to impact equals logic.