Ensure that user entries are inserted into the database in batch
These changes ensure that entries of type user are inserted as a
batch to the database rather than one-by-one updates. This speeds
up the load of data. This is possible by assuming that all entries
of type system appear before the first entry of type user.
Appending an entry of type system causes the write barrier to
break but appending an entry of type user does not. This
allows us to query the metadata index without decreasing load
performance.

Also, if the metadata index function is applied to a system entry
the IndexDriver should not try to index the entry. We should
change the way we specify IndexFunctions to include the entry
type it is relevant to.