Use latest Jetty and suppress remaining CVE
This commit upgrades Jetty from 9.4.2 to 9.4.6.v20170531 to fix an identified timing channel attack (CVE-2017-9735). We are upgrading the following libraries:

- org.eclipse.jetty:jetty-continuation,
- org.eclipse.jetty:jetty-http,
- org.eclipse.jetty:jetty-io,
- org.eclipse.jetty:jetty-security,
- org.eclipse.jetty:jetty-server,
- org.eclipse.jetty:jetty-servlet,
- org.eclipse.jetty:jetty-servlets,
- org.eclipse.jetty:jetty-util:",
- org.eclipse.jetty:jetty-webapp
- org.eclipse.jetty:jetty-xml

The following library is detailed as being vulnerable to the same CVE, but as of the time of writing, there are no updates which address it. We are therefore taking the decision to suppress this warning until a fix is made available.

- org.eclipse.jetty.toolchain.setuid:jetty-setuid-java:1.0.3