Cache metadata in Register object
We are assuming that no system entry will be added after the
user entry is appended to the register. This allows us to
(for now) cache the RegisterMetadata objects in memory after the
first time it is retrieved from the database, rather than going
to the database on each append-entry.

getTotalEntries() also does not require the all tables to be
flushed to the database and removing checkpoint() removes
soe unnecessary database queries.