Exclude system entries from particular index functions
With the addition of the new entry type property, we now need to ensure that we only run index functions on the appropriate type of data. For example, we should only run the `current-countries` index function when loading country data into a register, and similarly only run the `metadata` index function when loading system entries into a register.

If we do not check the entry type of incoming entries, we may incorrectly index system entries using any index functions available to the current register.