Read JDBC URL from environment when running on Cloud Foundry
Hack in support for deploying to Cloud Foundry/PaaS.

Ideally we wouldn't have to sniff the environment and parse out a JDBC
URL, however, the `DATABASE_URL` environment variable provided by PaaS
is a not a JDBC URL and not supported by the Java Postgres driver. We
will have to continue do this until a point where PaaS provides either:

- a JDBC URL as an environment variable.
- the individual host, port and database as environment variables so
we can construct a JDBC URL ourselves.

Currently the JDBC URL provided by PaaS doesn't specify `ssl=true` which
is required by PaaS. A ticket currently exists with PaaS to add this. We
add it manually here but can remove it once it's been added to the URL
the PaaS provides.