Create migration to populate index table with records
This commit will populate the index table with 'record' data for all of the records that currently reside in the register. It will does this by running through all of the entries of the register, ensuring that any changes of entry for a particular key are recorded so that the index can be used to track the history at a later date.