Change index functions to use a read only register
Currently, index functions take in their `execute` method a `Register`, which is the interface for a register which supports both reads and writes. Given that index functions only need to evaluate the current state of a register against some particular logic, there is no need for them to be given write access, so this commit changes all index functions to instead accept a `RegisterReadOnly`.