Explicitly specify the tag of the Docker image
This means that we can reason about which version of images are running.
We know that it'll be a specific (hopefully immutable) tag and not just
what happens to be "latest" at a given point. This also makes updating
images less tricky—you just update the tag, do a deploy and the new
image gets pulled in.