Specify an HTTP endpoint to use for health checks
It doesn't seem easy/possible to configure Cloud Foundry to use an HTTP
endpoint for its health check that isn't something served by a service
running on `$PORT`. This means the Dropwizard health check which runs on
`:8081/healthcheck` can't be used. Instead, we can use `/records` which
should exercise enough things about the application for us to be
reasonably convinced the application is "up".