Use CodeBuild to deploy to PaaS
This is very much a (mis)use of CodeBuild.

We currently deploy our code as part of a CodePipeline using CodeDeploy.
CodeDeploy works well with AWS-y things but doesn't have a way a neat
way to deploy to something externally like PaaS.

We can use CodeBuild to run a bunch of `cf` commands to push the `.jar`
to PaaS.

In order to do this our input to the CodeBuild project (the `.zip` that
Travis builds) has to contain a `buildspec.yml` and manifests the
describe the PaaS deployment--this is not ideal and our build artifact
has to know too much about its own deployment.

This feels very much like a stopgap solution to deploying to PaaS and
should be replaced if something more appropriate materializes.

There is a lot of duplication across manifests for different
environments. In the future I expect these manifests to define things
like the routes and domains associated with the app. Whilst they appear
similar now I think this is only accidental.

In order for this to work additional changes are required to
CodePipeline but the artifact must change first.