Run country locally instead of school-eng
The `./run-application` script used to run the school-eng alpha
register. This is now quite a large register and may move to beta
at some point. It therefore seems sensible to default to cloning
a smaller and more stable register like the beta country register.
The `json2rsf.py` script also required updating to be able to
specify the new system entries.