Use `autopilot` instead of `cf-blue-green-deploy`
We have had a few issues using `cf-blue-green-deploy`[1]; `autopilot`[2]
seems to work better for us. Problems we've seen include:

- Routes are not updated when making changes to `manifest.yaml`. They
appear to be just copied from the existing app. This makes adding new
registers a process that involves adding the routes outside of the
manifest.
- Deploys fail if there is not an app already running. This makes
rebuilding apps or creating apps for the first time awkward.

Apparent downsides to this include having to make sure that your
deployed application and manifest are similar[3]. This really should be
the case anyway as any deployment configuration changes should go
through the manifest.

[1] https://github.com/bluemixgaragelondon/cf-blue-green-deploy
[2] https://github.com/contraband/autopilot
[3] https://github.com/contraband/autopilot/blob/f602d9cbf609b5c99bee0caba452f7eb2bbf58ab/README.md#warning