Increase instance size to 4GB
Before moving to PaaS the size of each AWS EC2 instance was 4GB.
This was reduced to 1GB when we moved to PaaS but for no real
reason. Some instances are almost at their memory limit and
it's suspected that this could be causing some issues with
loading and querying data.