Fix spacing issue
This page resulted in output with unusual spacing before the comma and
full stop:

     The are 4,848 items , this sample shows only the first 100 .

This example text is from:
https://public-body.alpha.openregister.org/preview-records/json

The spacing appears because there is a newline between the comma/full
stop and the previous word in each case.

This commit fixes the issue by ensuring that there is no whitespace
between the previous word and the tag, and no whitespace between the
tag and the comma/full stop.