Fix Null Pointer Exception validating cardinality `n` curie fields
When calling `ItemValidator.validateFieldsValue` on a cardinality
`n` field there was a Null Pointer Exception being thrown by
`JsonNode.textValue()` because it was being called on `JsonArray`.
The offending logic was that checking whther a curie was valid. I
have moved this logic into a submethod that only operates on
individual components and never on a `JsonArray`.