RUNNING_LOCALLY.md should not suggest cloning a non-existent register
RUNNING_LOCALLY.md suggested cloning the school-type register
from the alpha phase, a register that no longer exists, which
causes errors when using the `run-application.sh` script. This
has now been updated not to specify the exact register, but
detailing where to look for registers that are possible to clone.