Refactor indexing code
There was some duplicate methods in `RecordIndex` and
`DerivationRecordIndex`. The first was specific to records and
the latter for more general indexes. Every method in `RecordIndex`
could be rewritten in terms of methods in terms of existing methods
in `DerivationRecordIndex`. This allowed the deletion of
`RecordIndex`. `DerivationRecordIndex` has also be renamed to `Index`.
Some general tidy up also.