Instruct travis to not timeout after 10 minutes
The travis build times out if it does not receive any output after
10 minutes. The registers build can take a long time sometimes due
to the OWASP dependency checker.
This change makes explicit that we use `./gradlew check`
(this is default anyway) and asks travis to wait for the check to
complete. We should look at how to plit up out or speed this step
up in a later PR if we can.