Correct error in RUNNING_LOCALLY.md
Before this change running
`ENVIRONMENT=discovery REGISTERS=uk ./run-application.sh` did not
work as per the instructions. This is because the first step to
load the local Register register failed. It failed because it was
trying to load data from the Register register in discovery and
the local register was checking for consistency against the beta
register. It would always fail when comparing the metadata being loaded
(with `phase=discovery`) against the existing metadata (with `phase=beta`).
The correct place to update `registersYaml` and `fieldsYaml` is therefore
`config.docker.basic.yaml` rather than `config.docker.register.yaml` as per
the instructions.