Calculate root hash for registers on start-up
This commit aims to warm-up the memoisation store for each register as soon as ORJ starts, by computing the root hash of each register. Through doing this, we will remove the burden of users needing to wait for this computation to take place before being able to download the RSF of a register, which should also have the effect of ensuring our caching layer is able to receive a response before timing out (using the current idle and keep-alive settings).