Upgrade jackson to 2.8.11
This commit upgrades Jackson from version 2.8.7 to 2.8.11. This is in response to the recently flagged vulnerability `CVE-2017-17485` (https://nvd.nist.gov/vuln/detail/CVE-2017-17485) which affects `jackson-databind` prior to 2.8.11, in addition to versions 2.9.0 up to and including 2.9.3. We could be affected by this vulnerability if the JSON that we read using `objectMapper.readValue` is crafted to exploit it, which could be achieved via loading registers with a specific payload.