Fix download link for entries of a record
This commit fixes the location of downloads for all formats when a user attempts to download the entries for a particular key. Previously, the download would point to `/entries` for the particular format, meaning that all entries would be downloaded, even when the user would be viewing only entries for some key. This commit now points downloads of this type to `/record/KEY/entries` for the particular format.