Download should use correct page info
This commit fixes a bug for both `/records` and `/entries` which currently leads you to download the wrong data, unless you're viewing page 1 with the default page size of 100 (or for entries, the default `start` of 1 and `limit` of 100).

With this commit, a user wishing to download the particular `/records` they're viewing, say page=3 with page-size=150, will be able to download exactly this data. Similarly, they can download the `/entries` they're viewing, from start=150 with limit=200.