WIP Record API key usage on requests
This commit is a work-in-progress commit, which uses a new `ApiKeyRecorderFilter` to intercept every request and insert an entry into a new `api_key_usage` table. This table is defined in the `public` schema and not on a per-schema basis, as we will want to ideally see all register usage in one place to make it easier to determine how users are using multiple registers. The table is by no means complete either - having something like a `session_id` might be useful.

To create the table, run the following:

```
create table if not exists public.api_key_usage
(
    api_key varchar not null,
    register varchar not null,
    path varchar not null,
    query_params varchar,
	timestamp integer
);
```

In terms of the `ApiKeyRecorderFilter`, it's necessary to establish a database connection but we can't inject an `ApiKeyDAO` directly into the filter. To get around this issue, the `DatabaseManager` reference from `RegisterApplication` has been essentially made a singleton, allowing us to inject `DatabaseManager` into the filter and use it's `getDbi` method to create an on-demand `ApiKeyDAO`. Again, this is not ideal on a per-request basis, but works for the proof-of-concept.

Finally, the main issue with logging API key usage inside of ORJ as far as I see it is we're not guaranteed to have all requests reach ORJ - if our caching layer returns cached responses to the user's request instead of going to origin.