Upgrade major version of Thymeleaf
This upgrade involved making changes to `ThymeleafContext` as it is
no longer possible to extend `WebContext`. `ThymeleafContext` now borrows
a lot of logic from Thymeleaf's `WebContext` and additionally creates
variables from properties in the relevant view, as before.
`ResourceResolver`s no longer exist in the new version. Instead, we use
a `FileTemplateResolver` for development (which means that it is possible
to make changes to templates whilst the app is running, as before) and
`ClassLoaderTemplateResolver` otherwise.