Iterate on entry table instead of index table to generate RSF
There was a recent refactor to use the indexing code for both
index and "register" actions, including creating RSF. The code
that reads the index table makes the assumption that there have
been no identical consecutive entries appended to the register.
This is not true for the existing government-service register and
therefore means that RSF is generated incorrectly for this register.

This is a temporary solution. Long term we should either fix the
history of the government-service register or no longer depend on this
rule to generate RSF.